requirements.txt
requests>=2.32
python-dotenv>=1.0
supabase>=2.4
flask>=3.0
